% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2GK.R
\name{D2Gs}
\alias{D2Gs}
\title{Convert pairwise distance matrix to Gower matrix}
\usage{
D2Gs(D)
}
\arguments{
\item{D}{pairwise distance matrix}
}
\value{
G centered Gower matrix
}
\description{
Adapted from the GRV R codes (Minas et al., 2013). Modified for better efficiency avoiding unnecessary matrix products.
Can show \eqn{G_{ij}= -0.5(D_{ij}^2-\bar{D^2}_{i,}-\bar{D^2}_{.,j}+\bar{D^2})}, where \eqn{\bar{D^2}_{i,},\bar{D^2}_{,j},\bar{D^2}} are
row, column and global averages. For symmetric distance matrix, \eqn{\bar{D^2}_{i,}=\bar{D^2}_{,i}}.
Note that G could be non-positive definite, and no further transformations are applied to G to make it non-negative definite.
}
\references{
Minas, C., Curry, E., Montana, G., 2013. A distance-based test of association between paired heterogeneous genomic data. Bioinformatics 29, 2555–2563. \url{http://wwwf.imperial.ac.uk/~gmontana/grv.html}

Guo,B. and Wu,B. (2017) On the fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12823.
}
