% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2GK.R
\name{D2Ks}
\alias{D2Ks}
\title{Convert pairwise distance matrix to kernels}
\usage{
D2Ks(D)
}
\arguments{
\item{D}{pairwise distance matrix}
}
\value{
\describe{
  \item{K}{ computed (non-negative) definite kernel matrix}
  \item{Kh}{ computed from eigen decomposition of K, such that \eqn{K=K_hK_h^T} }
  \item{Km}{ the effective rank of K }
}
}
\description{
Adapted from the MiRKAT package (Zhao et al., 2015). Modified for better efficiency avoiding unnecessary matrix products, and output
kernel and its eigen decomposition for followup analysis.
Can show \eqn{K_{ij}= -0.5(D_{ij}^2-\bar{D^2}_{i,}-\bar{D^2}_{.,j}+\bar{D^2})}, where \eqn{\bar{D^2}_{i,},\bar{D^2}_{,j},\bar{D^2}} are
row, column and global averages. For symmetric distance matrix, \eqn{\bar{D^2}_{i,}=\bar{D^2}_{,i}}.
}
\references{
Chen, J., Chen, W., Zhao, N., Wu, M.C., Schaid, D.J., 2016. Small Sample Kernel Association Tests for Human Genetic and Microbiome Association Studies. Genet. Epidemiol. 40, 5--19.

Zhao,N. et al. (2015) Testing in Microbiome-Profiling Studies with MiRKAT, the Microbiome Regression-Based Kernel Association Test. \emph{AJHG}, 96(5): 797–807.

Guo,B. and Wu,B. (2017) On the fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12823.
}
