% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRV_misc.R
\name{GRV.test}
\alias{GRV.test}
\title{GRV test for independence beteween paired heterogeneous genomic data from Minas et al. (2013)}
\usage{
GRV.test(DX, DY)
}
\arguments{
\item{DX}{distance matrix for a set of outcomes}

\item{DY}{distance matrix another set of outcomes}
}
\value{
test statistic and its associated p-value
}
\description{
GRV test for independence beteween paired heterogeneous genomic data from Minas et al. (2013)
}
\references{
Minas, C., Curry, E., Montana, G., 2013. A distance-based test of association between paired heterogeneous genomic data. Bioinformatics 29, 2555–2563.
}
