% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2GK.R
\name{K2C}
\alias{K2C}
\title{Create a centered Kernel matrix}
\usage{
K2C(K)
}
\arguments{
\item{K}{kernel matrix}
}
\value{
row and column centered kernel matrix
}
\description{
Given a symmetric kernel matrix K, we calculate \eqn{HKH}, where \eqn{H=I-11'/n} is a centering/projection matrix.
Mathematically it is equivalent to subtract the row and column means, which is computationally more efficient.
}
\references{
Guo,B. and Wu,B. (2017) On the fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12823.
}
