% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2GK.R
\name{KRV}
\alias{KRV}
\title{Compute the kernel independence test statistic between two kernels efficiently}
\usage{
KRV(KX, KY)
}
\arguments{
\item{KX}{centered symmetric kernel matrix for a set of outcomes}

\item{KY}{centered symmetric kernel matrix for another set of outcomes}
}
\value{
kernel test statistic
}
\description{
For symmetric kernel matrices, we can efficiently compute the kernel test statistic as a form of correlation,
\eqn{(\sum_{i,j} X_{ij}Y_{ij})/\sqrt{(\sum_{ij}X_{ij}^2)(\sum_{ij}Y_{ij}^2)} }.
}
\references{
Minas, C., Curry, E., Montana, G., 2013. A distance-based test of association between paired heterogeneous genomic data. Bioinformatics 29, 2555–2563.

Zhan, X., Plantinga, A., Zhao, N., Wu, M.C. (2017). A fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12684.

Guo,B. and Wu,B. (2017) On the fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12823.
}
