% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2GK.R
\name{KRVtest}
\alias{KRVtest}
\title{Revised implementation of KRV test for independence between paired data}
\usage{
KRVtest(KX, KY, kernel = FALSE)
}
\arguments{
\item{KX}{distance matrix or the computed kernel matrix for a set of outcomes}

\item{KY}{distance matrix or the computed kernel matrix for another set of outcomes}

\item{kernel}{indicator of whether the input is kernel matrix computed using D2Ks() and K2C().}
}
\value{
test statistic and its p-value
}
\description{
Revised implementation of KRV test for independence between paired data
}
\references{
Minas, C., Curry, E., Montana, G., 2013. A distance-based test of association between paired heterogeneous genomic data. Bioinformatics 29, 2555–2563.

Zhan, X., Plantinga, A., Zhao, N., Wu, M.C. (2017). A fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12684.

Guo,B. and Wu,B. (2017) On the fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12823.
}
