% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_misc.R
\name{RVgamma}
\alias{RVgamma}
\title{Approximately model RV statistic Q with a shifted gamma distribution Gamma(k,\eqn{\theta})}
\usage{
RVgamma(mu, sigma, s, r)
}
\arguments{
\item{mu}{mean of permuted Q}

\item{sigma}{standard error of permuted Q}

\item{s}{skewness: \eqn{E(Q-\mu)^3/\sigma^3}}

\item{r}{kurtosis: \eqn{E(Q-\mu)^4/\sigma^4}}
}
\value{
\describe{
  \item{par3}{estimated (k,\eqn{\theta},\eqn{\lambda}) when matching skewness}
  \item{par4}{estimated (k,\eqn{\theta},\eqn{\lambda}) when matching kurtosis}
}
}
\description{
The idea is to model \eqn{Q-\lambda} with Gamma(k,\eqn{\theta}). We choose (k,\eqn{\theta},\eqn{\lambda}) to match
the mean, variance and skewness or kurtosis of Q (all computed under permutation distribution).
The Gamma PDF \eqn{\frac{1}{\Gamma(k)\theta^k}x^{k-1}e^{-x/\theta}}, mean \eqn{k\theta}, variance \eqn{k\theta^2},
skewness \eqn{2/\sqrt{k}}, kurtosis \eqn{6/k+3}.
}
\references{
Minas, C., Curry, E., Montana, G., 2013. A distance-based test of association between paired heterogeneous genomic data. Bioinformatics 29, 2555–2563.

Zhan, X., Plantinga, A., Zhao, N., Wu, M.C. (2017). A fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12684.

Guo,B. and Wu,B. (2017) On the fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12823.
}
