% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_misc.R
\name{RVparam}
\alias{RVparam}
\title{Calculate the first four permutation moments: mean, variance, skewness and kurtosis for RV statistic}
\usage{
RVparam(KX, KY)
}
\arguments{
\item{KX}{the first kernel matrix}

\item{KY}{the second kernel matrix}
}
\value{
a vector of mean, standard error, skewness, and kurtosis
}
\description{
Calculate the first four permutation moments: mean, variance, skewness and kurtosis for RV statistic
}
\references{
Kazi-Aoual, F., Hitier, S., Sabatier, R., Lebreton, J.-D. (1995). Refined Approximations to Permutation Tests for Multivariate Inference. Comput. Stat. Data Anal. 20, 643–656.

Guo,B. and Wu,B. (2017) On the fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12823.
}
