% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_misc.R
\name{RVtest}
\alias{RVtest}
\title{Unified interface to conduct general RV test using various Gamma approximations}
\usage{
RVtest(KX, KY)
}
\arguments{
\item{KX}{the first centered kernel/Gower matrix}

\item{KY}{the second centered kernel/Gower matrix}
}
\value{
a vector of three p-values: Gamma approx by matching skewness and kurtosis respectively, plus the permutation p-value
}
\description{
Given centered kernel/Gower matrices, test their independence.
}
\references{
Guo,B. and Wu,B. (2017) On the fast small-sample kernel independence test for microbiome community-level association analysis. Biometrics, doi:10.1111/biom.12823.
}
